Start of attempt to support writing to the server:
1. Implement a set of class-based wrappers which can be declared
on the stack, and thus automatically deal with resource deallocation
of all the myriad error paths. These could be used throughout the
code to the same end.

2. Implement comprehensive error reporting logic in support of the
above.

3. Flesh out the calendar change path, but leave it disabled. The logic
in mapiconnector2.cpp exposes a bug in libmapi where SetProps asserts in
cast_mapi_SPropValue() because it omits support for PT_MV_CLSID.